+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 
 Changes:
 
+v0.0.2
+  - allow proxy any income trafic without method knowlage
+  - fix name peak_first to pick_first
+
 v0.0.1
   - added simple example
   - added documentation